# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: KIWAME API REST
  description: API for the management of KIWAME 
...
securityDefinitions:
  ApiKey:
    type: "apiKey"
    name: "x-api-key"
    in: "header"
    description: "Security sheme migrated from security policy"
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
     For accessing the API a valid JWT token must be passed in all the queries in
     the 'Authorization' header.


     A valid JWT token is generated by the API and retourned as answer of a call
     to the route /login giving a valid user & password.


     The following syntax must be used in the 'Authorization' header :

       Bearer xxxxxx.yyyyyyy.zzzzzz


# during dev, should point to your local machine
#host: 80.58.61.69  #PRODUCCIÓN
#host: 138.4.47.33:8163  #UPM DEV
#host: localhost:8090  #LOCAL DEV
# basePath prefixes all resource paths
basePath: / #dds
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /rule/types:
    x-swagger-router-controller: rule
    get:
      tags:
        - rules
      summary: Retrieve all rule tpyes
      description: Retreive all rules types
      operationId: getRuleTypes
      security:
        - Bearer: []
        - ApiKey: []
      x-security-scopes:
        - profesional
        - admin 
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - description: Provide a header to specify the language.
          in: header
          name: Accept-Language
          required: false
          type: string
      responses:
        "200":
          description: rules retrieved
          schema:
            $ref: "#/definitions/RulesTypeResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Rules types Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /rule/conditionTypes:
    x-swagger-router-controller: rule
    get:
      tags:
        - rules
      summary: Retrieve condition tpyes
      description: Retreive all condition
      operationId: getConditionTypes
      security:
        - Bearer: []
        - ApiKey: []
      x-security-scopes:
        - profesional
        - admin 
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - description: Provide a header to specify the language.
          in: header
          name: Accept-Language
          required: false
          type: string
      responses:
        "200":
          description: Conditions retrieved
          schema:
            $ref: "#/definitions/ConditionsTypeResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Conditions types Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
  /rule:
    x-swagger-router-controller: rule
    post:
      tags:
        - rules
      operationId: newRule
      summary: post a new rule
      description: Upload a new rule
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
      produces: 
        - application/json
      parameters:
        - name:  rule
          in: body
          description: The rule JSON you want to post
          schema:
            $ref: "#/definitions/RuleObject"
          required: true

      responses:
        "201":
          description: "Success"
          schema:
            $ref: "#/definitions/ServerResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /rule/default:
    x-swagger-router-controller: rule
    post:
      tags:
        - rules
      operationId: defaultRules
      summary: create default rules
      description: create the default rules for a user
      security:
        - Bearer: []
      x-security-scopes:
        - patient
      consumes: 
        - application/json
        - application/octet-stream
      produces: 
        - application/json
      responses:
        "201":
          description: "Success"
          schema:
            $ref: "#/definitions/ServerResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: "User not found or missing arguments"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse" 
  /rule/{id}:
    x-swagger-router-controller: rule
    get:
      tags:
        - rules
      summary: Toogle enabled 
      description: toogle enable/disabled rule
      operationId: findById
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to toogle
          type: string
          required: true
      responses:
        "200":
          description: rule retrieves
          schema:
            $ref: "#/definitions/RuleResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Rule Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    put:
      tags:
        - rules
      summary: Toogle enabled 
      description: toogle enable/disabled rule
      operationId: toogleEnabled
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to toogle
          type: string
          required: true
      responses:
        "200":
          description: rule Enabled/Disabled
          schema:
            $ref: "#/definitions/ServerResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Rule Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    delete:
      tags:
        - rules
      summary: Delete rule
      description: Delete rule
      operationId: deleteRule
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to delete
          type: string
          required: true
      responses:
        "200":
          description: Rule removed
          schema:
            $ref: "#/definitions/ServerResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Profile Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
definitions:
...
  RuleObject:
    required:
      - type
      - condition
      - value
      - notificationType
    properties:
      patient:
        type: string
        example: nickname
      type:
        type: number
        example: 1
      measureType:
        type: number
        example: 2
      nutrientType:
        type: number
        example: 1
      condition:
        type: number
        example: 1
      value:
        type: number
        example: 10
      period:
        type: number
        minimum: 1
        maximum: 30
        example: 1
      notificationType:
        type: number
        example: 1
  RuleRetrievedObject:
    required:
      - type
      - condition
      - value
      - notificationType
    properties:
      id:
        type: string
      patient:
        type: string
        example: nickname
      type:
        type: number
        example: 1
      measureType:
        type: number
        example: 2
      nutrientType:
        type: number
        example: 1
      condition:
        type: number
        example: 1
      value:
        type: number
        example: 10
      period:
        type: number
        minimum: 1
        maximum: 30
        example: 1
      notificationType:
        type: number
        example: 1
      enabled:
        type: boolean
        example: true
      created:
        type: string
        format: date-time
        example: "2016-06-14T10:24:11.387Z"
      updated:
        type: string
        format: date-time
        example: "2016-06-14T10:24:11.387Z"
  RuleTypeObject:
    required:
      - type
      - name
      - description
    properties:
      type:
        type: number
        example: 1
      name:
        type: string
        example: rule name
      description:
        type: string
        example: description
...
  ConditionTypeObject:
    required:
      - type
      - description
    properties:
      type:
        type: number
        example: 1
      description:
        type: string
        example: description
...
  NotFoundResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Error: Resource Not found"
      statusCode:
        type: number
        example: 404
...
  RuleResponse:
    required:
      - message
      - statusCode
      - rule
    properties:
      message:
        type: string
        example: "success"
      statusCode:
        type: number
        example: 200
      rule:
          $ref: "#/definitions/RuleRetrievedObject"
  ServerResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Success"
      statusCode:
        type: number
        example: 200
...
  BadRequestResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Bad Request"
      statusCode:
        type: number
        example: 400
  AccessDeniedResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Access Denied"
      statusCode:
        type: number
        example: 403
...
  RulesTypeResponse:
    required:
      - message
      - statusCode
      - rules
    properties:
      message:
        type: string
        example: "success"
      statusCode:
        type: number
        example: 200
      rules:
          type: array
          items: 
            $ref: "#/definitions/RuleTypeObject"
...
  ConditionsTypeResponse:
    required:
      - message
      - statusCode
      - conditions
    properties:
      message:
        type: string
        example: "success"
      statusCode:
        type: number
        example: 200
      conditions:
          type: array
          items: 
            $ref: "#/definitions/ConditionTypeObject"
...
  ServerErrorResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Error: Internal Server Error"
      statusCode:
        type: number
        example: 500